refactor: ディレクトリ構造を整理
