+++ endpoints/guilds/router.py
@@ -1,10 +1,10 @@
 import asyncpg
 from fastapi import APIRouter, Security
 
+from core.auth import verify_bearer_token
+from db.connection import get_connection_pool
 from endpoints.guilds import dict_api, settings_api, character_usage_api, trusted_roles_api, connection_command_api, \
     message_link_expand_pref_api, connection_states_api, subscriptions_api
-from utils.auth import verify_bearer_token
-from utils.db_connection import get_connection_pool
 
 router = APIRouter()
 router.include_router(dict_api.router)
