+++ endpoints/guilds/trusted_roles_api.py
@@ -6,8 +6,8 @@ import asyncpg
 from fastapi import APIRouter, Security
 from pydantic import BaseModel
 
-from utils.auth import verify_all_tokens
-from utils.db_connection import get_connection_pool
+from core.auth import verify_all_tokens
+from db.connection import get_connection_pool
 
 router = APIRouter()
