+++ endpoints/shards/router.py
@@ -3,9 +3,9 @@ from typing import Literal, Optional
 import asyncpg
 from fastapi import APIRouter, Security
 
+from core.auth import verify_session
+from db.connection import get_connection_pool
 from endpoints.shards import release_api, assign_api, connection_commands_api, metrics_api
-from utils.auth import verify_session
-from utils.db_connection import get_connection_pool
 
 router = APIRouter()
 router.include_router(assign_api.router)
